Fix build errors: TopControls file mismatch and Kotlin 2.3 compatibility
- Resolved persistent `TopControls.kt` compilation errors by ensuring the correct stub implementation is used in the CI environment (moved from `ui/TopControls.kt` to `ui/composables/TopControls.kt`).
- Fixed Kotlin 2.3.0/Hilt metadata conflict by adding `ksp(libs.kotlin.metadata.jvm)` to build dependencies.
- Provided missing `Gson` binding in `AppModule` to fix Hilt injection failure.
- Updated `MainLayout` and `HaterScreen` to correctly cast `TableSize.feet` to Float.
- Fixed `BallTextRenderer` instantiation logic.
- Cleaned up legacy AR code and `MainScreen.kt`.
- Verified build and tests pass.

+++ app/src/main/java/com/hereliesaz/cuedetat/ui/hatemode/HaterScreen.kt
@@ -144,9 +144,9 @@ fun HaterScreen(
             areHelpersVisible = uiState.areHelpersVisible,
             experienceMode = uiState.experienceMode,
             isTableVisible = uiState.table.isVisible,
-            tableSizeFeet = uiState.table.size.feet,
+            tableSizeFeet = uiState.table.size.feet.toFloat(),
             isBeginnerViewLocked = uiState.isBeginnerViewLocked,
-            targetBallDistance = uiState.targetBallDistance,
+            targetBallDistance = uiState.targetBallDistance.toFloat(),
             distanceUnit = uiState.distanceUnit,
             onCycleTableSize = { onEvent(MainScreenEvent.CycleTableSize) },
             onMenuClick = { onEvent(MainScreenEvent.ToggleNavigationRail) }
